+++ src/constants/cardNumber.js
@@ -2,3 +2,4 @@ export const FIRST_NUMBER = "firstNumber";
 export const SECOND_NUMBER = "secondNumber";
 export const THIRD_NUMBER = "thirdNumber";
 export const FOURTH_NUMBER = "fourthNumber";
+export const CARD_LIST = "cardList";
