+++ src/pages/CardList.jsx
@@ -1,16 +1,15 @@
 import Card from "../components/atomic-design-pattern/molecule/Card";
 import CardBox from "../components/CardBox";
-import {
-  CardStorage,
-  getLocalStorageKeyByCardNumber,
-} from "../util/cardStorage";
+import { getLocalStorageKeyByCardNumber } from "../util/cardStorage";
 import { MONTH, YEAR } from "../constants/expirationDate";
 import { useContext } from "react";
 import { CardContext } from "../../providers/CardState/CardStateProvider";
+import { useLocalStorage } from "../hook/useLocalStorage";
 
 export default function CardList({ onNext, goToCompletePage }) {
+  const { getCardInfoList } = useLocalStorage();
   const { setCardState } = useContext(CardContext);
-  const cardInfoList = CardStorage.getCardInfoList();
+  const cardInfoList = getCardInfoList();
   const onClickCard = (cardInfo) => {
     setCardState(cardInfo);
     goToCompletePage();
